+++ tests/rollup.config.js
@@ -1,4 +1,3 @@
-import babel from 'rollup-plugin-babel';
 import multiEntry from 'rollup-plugin-multi-entry';
 
 export default {
@@ -8,13 +7,6 @@ export default {
     format: 'cjs',
     sourcemap: true
   },
-  external: ['ava'],
-  plugins: [
-    babel({
-      babelrc: false,
-      presets: ['env'],
-      exclude: 'node_modules/**'
-    }),
-    multiEntry()
-  ]
+  external: ['ava', 'kronos-service', 'uti'],
+  plugins: [multiEntry()]
 };
